Working trait selector for actor traits
