+++ src/module/apps/trait-selector.js
@@ -10,20 +10,11 @@ export class TraitSelectorSFRPG extends FormApplication {
         options.id = "trait-selector";
         options.classes = ["sfrpg"];
         options.title = "Trait Selection";
-        options.template = "systems/sfrpg/templates/apps/trait-selector.hbs";
+        options.template = "systems/sfrpg/templates/apps/trait-selector-new.hbs";
         options.width = 480;
         options.height = "auto";
 
         return options;
-    }
-
-    /**
-     * Return a reference to the target attribute
-     *
-     * @type {String}
-     */
-    get attribute() {
-        return this.options.name;
     }
 
     /**
@@ -33,104 +24,109 @@ export class TraitSelectorSFRPG extends FormApplication {
      */
     getData() {
         console.log(this);
-        const attr = getProperty(this.object, this.attribute);
-        if (typeof attr.value === "string") {
-            attr.value = this.constructor._backCompat(attr.value, this.options.choices);
+
+        // Initialize variables for easy access
+        const dataFormat = this.options.format;
+        const dataLocation = this.options.location;
+        const traitData = getProperty(this.object, dataLocation);
+
+        // Choose the appropriate data parser to create the form
+        switch (dataFormat) {
+            case "actorTraits":
+                return this._getActorTraitChoices(traitData);
+
+            case "weaponProperties":
+                console.log('Selecting Weapon Properties');
+                return true;
+
+            default:
+                console.log(`dataFormat ${dataFormat} not found`);
+                break;
         }
-
-        const choices = duplicate(this.options.choices);
-        const isEnergyResistance = this.attribute === "data.traits.dr";
-        if (!isEnergyResistance) {
-            for (const [k, v] of Object.entries(choices)) {
-                choices[k] = {
-                    label: v,
-                    chosen: attr.value.includes(k)
-                };
-            }
-        } else {
-            for (const [k, v] of Object.entries(choices)) {
-                choices[k] = {
-                    label: v,
-                    chosen: false,
-                    resistanceValue: 0
-                };
-            }
-
-            for (const value of attr.value) {
-                for (const [type, resistance] of Object.entries(value)) {
-                    choices[type].chosen = true;
-                    choices[type].resistanceValue = resistance;
-                }
-            }
-        }
-
-        this.searchTerm = '';
-
-        return {
-            choices: choices,
-            custom: attr.custom,
-            isEnergyResistance
-        };
     }
 
     /**
-     * Support backwards compatability for old-style string separated traits
-     *
-     * @param {String} current The current value
-     * @param {Array} choices The choices
-     * @returns {Array}
-     * @private
-     */
-    static _backCompat(current, choices) {
-        if (!current || current.length === 0) return [];
-        current = current.split(/[\s,]/).filter(t => !!t);
-        return current.map(val => {
-            for (const [k, v] of Object.entries(choices)) {
-                if (val === v) return k;
-            }
-            return null;
-        }).filter(val => !!val);
-    }
-
-    /**
-     * Update the Actor object with new trait data processed from the form
+     * Choose the appropriate update method for updating the data
      *
      * @param {Event} event The event that triggers the update
      * @param {Object} formData The data from the form
      * @private
      */
     _updateObject(event, formData) {
-        let choices = [];
 
-        if (this.attribute !== "system.traits.dr") {
-            for (const [k, v] of Object.entries(formData)) {
-                if ((k !== 'custom') && v) choices.push(k);
+        // Get the data format
+        const dataFormat = this.options.format;
+
+        switch (dataFormat) {
+            case 'actorTraits':
+                console.log('Setting Actor Traits');
+                this._setActorTraits(formData);
+                break;
+
+            default:
+                console.log('Setting nothing.');
+                break;
+        }
+    }
+
+    /**
+     * Parses Actor Trait data into a format that the form can accept
+     *
+     * @param {Object} traitData The data from the actor to parse
+     * @returns {Object}
+     */
+    _getActorTraitChoices(traitData) {
+
+        // create the array of choices
+        const choices = duplicate(this.options.choices);
+        console.log(choices, traitData);
+
+        for (const [k, v] of Object.entries(choices)) {
+            choices[k] = {
+                label: v,
+                isSelected: traitData.value.includes(k)
+            };
+        }
+        return {
+            choices: choices,
+            custom: traitData.custom
+        };
+    }
+
+    /**
+     * Update an Actor trait data processed from the form
+     *
+     * @param {Object} formData The data from the form
+     * @private
+     */
+    _setActorTraits(formData) {
+
+        // get a list of valid choices and initialize array
+        const validChoices = Object.keys(this.options.choices);
+        const selected = [];
+
+        // Push custom values first, then others, ignoring options not in the list of choices
+        // key is the specific language, proficiency, etc.
+        // value is true or false, or the name of a custom trait
+        for (const [key, value] of Object.entries(formData)) {
+            if (validChoices.includes(key)) {
+                if (value) selected.push(key);
             }
-        } else {
-            let resistances = Object.entries(formData).filter(e => e[0].startsWith("er"));
-            resistances = resistances.reduce((obj, entry) => {
-                const [type, i] = entry[0].split('.').slice(1);
-
-                if (!obj[type]) obj[type] = {};
-                obj[type][i] = entry[1];
-
-                return obj;
-            }, {});
-
-            choices = Object.entries(resistances).filter(e => e[1][0])
-                .reduce((arr, resistance) => {
-                    arr.push({[resistance[0]]: resistance[1][1]});
-
-                    return arr;
-                }, []);
         }
 
+        // update parent object with the appropriate data
         this.object.update({
-            [`${this.attribute}.value`]: choices,
-            [`${this.attribute}.custom`]: formData.custom
+            [`${this.options.location}.value`]: selected,
+            [`${this.options.location}.custom`]: formData.custom
         });
     }
 
+    /**
+     * Filter the displayed traits based on search input
+     *
+     * @param {Array} li An array of all the traits displayed on the form
+     * @private
+     */
     async filterTraits(li) {
         li.hide();
 
@@ -141,6 +137,12 @@ export class TraitSelectorSFRPG extends FormApplication {
         }
     }
 
+    /**
+     * Check if a displayed item matches the search string
+     *
+     * @param {String} trait The localized trait name
+     * @private
+     */
     searchMatch(trait) {
         const searchTerm = this.searchTerm;
 
@@ -165,8 +167,6 @@ export class TraitSelectorSFRPG extends FormApplication {
             this.searchTerm = ev.target.value;
             this.filterTraits(html.find('li'));
         });
-
-        // on Enter key press, filter based on search
 
     }
 }
