+++ src/templates/apps/trait-selector.hbs
@@ -7,27 +7,14 @@
 
     {{!-- Trait List --}}
     <ul class="trait-list">
-    {{#unless isEnergyResistance}}
         {{#each choices as |choice key|}}
-        <li class="trait">
-            <label class="checkbox">
-                <input type="checkbox" name="{{key}}" data-dtype="Boolean" {{checked choice.chosen}}>
-                {{choice.label}}
-            </label>
-        </li>
+            <li class="trait">
+                <label class="checkbox">
+                    <input type="checkbox" name="{{key}}" data-dtype="Boolean" {{checked choice.isSelected}}>
+                    {{choice.label}}
+                </label>
+            </li>
         {{/each}}
-    {{else}}
-        {{#each choices as |choice key|}}
-        <li class="flexrow energy-resistance trait">
-            <label class="checkbox">
-                <input type="checkbox" name="er.{{key}}.0" data-dtype="Boolean" {{checked choice.chosen}}>
-                {{choice.label}}
-            </label>
-
-            <input type="text" name="er.{{key}}.1" data-dtype="Number" value="{{choice.resistanceValue}}" />
-        </li>
-        {{/each}}
-    {{/unless}}
     </ul>
 
     {{!-- Special Traits --}}
